pyproject.toml
[tool.poetry]
name = "referal-tg-bot"
version = "0.1.0"
description = ""
authors = ["matvey <matgr2n@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.3.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
